Working version open loop - found bug where encoder i2c read blocks the serial comms

+++ src/SerialManager/SerialManager.h
@@ -125,7 +125,7 @@ private:
     // CRC //
     //////////////////
 
-    uint8_t _computeCRC(const uint8_t* data, uint8_t length);
+    uint8_t _computeCRC(const uint8_t* data, size_t length);
 
     //////////////////////////
     // Internal State //
